requirements.txt
Flask==2.3.3
pdfplumber==0.9.0
python-docx==0.8.11
google-generativeai
groq
python-dotenv
scikit-learn
nltk
numpy
gunicorn